product-routes.js
Ive completed the routes for the products using the same restful route syntax as categories. However this code already had some routes completed due to the starter code and that code does not use restful routes. Im still trying to decided whether I change the starter code to match my own syntax or leave it be.