File Structure and Small Fixes
I removed the Develop folder and placed everything into the main repository folder as the starter code was made to be in that format. I also made a couple of syntax fixes. Now npm run seed works but is showing that I need a password. Currently looking to solve this.